Weight adjustments are now saved, weights are integers

+++ src/main/java/net/spacerulerwill/skygrid/util/BlockWeight.java
@@ -5,11 +5,11 @@ import com.mojang.serialization.codecs.RecordCodecBuilder;
 import net.minecraft.block.Block;
 import net.minecraft.registry.Registries;
 
-public record BlockWeight(Block block, double weight) {
+public record BlockWeight(Block block, int weight) {
     public static final Codec<BlockWeight> CODEC = RecordCodecBuilder.create(
             instance -> instance.group(
                     Registries.BLOCK.getCodec().fieldOf("block").forGetter(BlockWeight::block),
-                    Codec.DOUBLE.fieldOf("weight").forGetter(BlockWeight::weight)
+                    Codec.INT.fieldOf("weight").forGetter(BlockWeight::weight)
             ).apply(instance, BlockWeight::new)
     );
 }
